+++ mtgjson5/classes/mtgjson_game_formats.py
@@ -3,7 +3,7 @@ MTGJSON Singular Card.GameFormats Object
 """
 from typing import List
 
-from mtgjson5.utils import to_camel_case
+from ..utils import to_camel_case
 
 
 class MtgjsonGameFormatsObject:
