+++ mtgjson5/providers/scryfall/set_language_detector.py
@@ -9,7 +9,6 @@ from singleton_decorator import singleton
 from ...constants import LANGUAGE_MAP
 from ...providers.abstract import AbstractProvider
 from ...providers.scryfall import sf_utils
-from ...utils import retryable_session
 
 LOGGER = logging.getLogger(__name__)
