+++ mtgjson5/utils.py
@@ -2,10 +2,7 @@
 MTGJSON simple utilities
 """
 import collections
-import datetime
-import functools
 import hashlib
-import inspect
 import itertools
 import json
 import logging
@@ -17,11 +14,9 @@ from typing import Any, Callable, Dict, Iterator, List, Optional, Set, Tuple, Un
 import gevent.pool
 import requests
 import requests.adapters
-import requests_cache
-import urllib3
 
-from mtgjson5 import constants
-from mtgjson5.mtgjson_config import MtgjsonConfig
+from . import constants
+from .mtgjson_config import MtgjsonConfig
 
 LOGGER = logging.getLogger(__name__)
 
@@ -99,48 +94,6 @@ def parse_magic_rules_subset(
     valid_line_segments = "\n".join(magic_rules.splitlines())
 
     return valid_line_segments
-
-
-def retryable_session(
-    retries: int = 8,
-) -> Union[requests.Session, requests_cache.CachedSession]:
-    """
-    Session with requests to allow for re-attempts at downloading missing data
-    :param retries: How many retries to attempt
-    :return: Session that does the downloading
-    """
-    session: Union[requests.Session, requests_cache.CachedSession]
-
-    if MtgjsonConfig().use_cache:
-        stack = inspect.stack()
-        calling_class = stack[1][0].f_locals["self"].__class__.__name__
-        session = requests_cache.CachedSession(
-            cache_name=str(constants.CACHE_PATH.joinpath(calling_class)),
-            expire_after=datetime.timedelta(days=1),
-            stale_if_error=True,
-        )
-    else:
-        session = requests.Session()
-
-    retry = urllib3.util.retry.Retry(
-        total=retries,
-        read=retries,
-        connect=retries,
-        backoff_factor=0.3,
-        status_forcelist=(500, 502, 504),
-    )
-
-    adapter = requests.adapters.HTTPAdapter(max_retries=retry)
-    session.mount("http://", adapter)
-    session.mount("https://", adapter)
-    session.request = functools.partial(session.request, timeout=5)  # type: ignore
-
-    session.headers.update(
-        {
-            "User-Agent": "Mozilla/5.0 (X11; Linux x86_64; +https://www.mtgjson.com) Gecko/20100101 Firefox/120.0"
-        }
-    )
-    return session
 
 
 def parallel_call(
